Moved input validation files to input_validation folder to better organize the project
